Stack mask preview vertically, group screens into feature folders, and animate loading overlays
