+++ app/src/main/java/com/example/capilux/screen/mask/MaskPreviewScreen.kt
@@ -1,4 +1,4 @@
-package com.example.capilux.screen
+package com.example.capilux.screen.mask
 
 import android.net.Uri
 import android.graphics.BitmapFactory
@@ -14,6 +14,9 @@ import androidx.compose.ui.graphics.asImageBitmap
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavHostController
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.CardDefaults
+import androidx.compose.foundation.BorderStroke
 import com.example.capilux.ui.theme.backgroundGradient
 import java.io.File
 
@@ -45,39 +48,49 @@ fun MaskPreviewScreen(
 
         Spacer(modifier = Modifier.height(24.dp))
 
-        Row(
+        Column(
             modifier = Modifier.fillMaxWidth(),
-            horizontalArrangement = Arrangement.SpaceEvenly
+            horizontalAlignment = Alignment.CenterHorizontally
         ) {
-            // Foto original
+            // Foto original arriba
             Column(horizontalAlignment = Alignment.CenterHorizontally) {
                 Text("Original", color = Color.White)
                 if (originalFile.exists()) {
                     val bitmap = BitmapFactory.decodeFile(originalFile.absolutePath)
-                    Image(
-                        bitmap = bitmap.asImageBitmap(),
-                        contentDescription = "Imagen original",
-                        modifier = Modifier
-                            .size(160.dp)
-                            .padding(8.dp)
-                    )
+                    Card(
+                        shape = RoundedCornerShape(12.dp),
+                        border = BorderStroke(1.dp, Color.White.copy(alpha = 0.3f)),
+                        colors = CardDefaults.cardColors(containerColor = Color.White.copy(alpha = 0.1f))
+                    ) {
+                        Image(
+                            bitmap = bitmap.asImageBitmap(),
+                            contentDescription = "Imagen original",
+                            modifier = Modifier.size(160.dp)
+                        )
+                    }
                 } else {
                     Text("❌ Sin imagen", color = Color.Red)
                 }
             }
 
-            // Máscara generada
+            Spacer(modifier = Modifier.height(24.dp))
+
+            // Máscara generada abajo
             Column(horizontalAlignment = Alignment.CenterHorizontally) {
                 Text("Máscara", color = Color.White)
                 if (maskFile.exists()) {
                     val bitmap = BitmapFactory.decodeFile(maskFile.absolutePath)
-                    Image(
-                        bitmap = bitmap.asImageBitmap(),
-                        contentDescription = "Máscara",
-                        modifier = Modifier
-                            .size(160.dp)
-                            .padding(8.dp)
-                    )
+                    Card(
+                        shape = RoundedCornerShape(12.dp),
+                        border = BorderStroke(1.dp, Color.White.copy(alpha = 0.3f)),
+                        colors = CardDefaults.cardColors(containerColor = Color.White.copy(alpha = 0.1f))
+                    ) {
+                        Image(
+                            bitmap = bitmap.asImageBitmap(),
+                            contentDescription = "Máscara",
+                            modifier = Modifier.size(160.dp)
+                        )
+                    }
                 } else {
                     Text("❌ No generada", color = Color.Red)
                 }
